pyproject.toml
[project]
name = "migru"
version = "1.0.0"
description = "A warm, AI-powered companion for migraine and stress support with real-time pattern detection"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Ash-Blanc", email = "angelash18092007@gmail.com"}
]
keywords = ["wellness", "migraine", "stress", "ai", "companion", "real-time-analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

# Production dependencies only
dependencies = [
    "agno>=2.3.24",              # AI agent framework
    "redis>=5.0.0",              # Local data storage
    "python-dotenv>=1.0.0",      # Environment configuration
    "rich>=13.0.0",              # Beautiful CLI
    "pathway>=0.28.0",           # Real-time streaming analytics
    "youtube-transcript-api>=1.2.3",
    "cerebras-cloud-sdk>=1.64.1",
    "ddgs>=9.10.0",
    "firecrawl-py>=4.12.0",
    "mistralai>=1.10.0",
    "openai>=2.15.0",
    "prompt_toolkit>=3.0.0",
    "psutil>=7.2.1",
    "pyarrow>=18.1.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",              # HTTP client for local LLM APIs
]

[project.optional-dependencies]
# Development dependencies (not needed for end users)
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.3.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "psutil>=5.9.0",
]

# Local LLM support (for enhanced privacy)
local = [
    "ollama>=0.1.0",                     # Ollama client (optional)
    "llama-cpp-python>=0.2.0",           # Direct llama.cpp support (optional)
]

[project.scripts]
migru = "app.main:app"

[project.urls]
Homepage = "https://github.com/Ash-Blanc/migru"
Repository = "https://github.com/Ash-Blanc/migru"
Documentation = "https://github.com/Ash-Blanc/migru#readme"
"Bug Tracker" = "https://github.com/Ash-Blanc/migru/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config -x --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "performance: marks tests as performance tests",
    "cli: marks tests as CLI tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/.mypy_cache/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "except ImportError:",
    "raise",
    "return NotImplemented",
]

# Ruff linting configuration
[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # flake8-bandit
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "S101",  # assert used
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "F811", "F841"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
force-single-line = true
known-first-party = ["app"]
known-third-party = ["pytest", "uv", "ruff"]

# MyPy type checking configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    "agno.*",
    "redis.*",
    "ddgs.*",
    "firecrawl.*",
    "cerebras_cloud_sdk.*",
    "pathway.*",
    "pytest.*",
    "uv.*",
    "ruff.*",
]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
